Check presence of `params[:group]` when deactivating group rollout
This avoids an error (NoMethodError: undefined method `to_sym' for
<ActionController::Parameters {} permitted:
false>:ActionController::Parameters) that otherwise occurs when
deactivating a rollout for a group, if the request is made in such a way
as to trigger [parameter wrapping][1] (e.g. if the request is made with
`Content-Type: application/json` and the application is configured with
`wrap_parameters format: [:json]`).

[1]: https://api.rubyonrails.org/v5.2.3/classes/ActionController/ParamsWrapper.html